sili_telegram_bot: Move response types to models
  Same as with Responses
